# backend/internal/http/docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenShop API
  version: 1.0.0
  description: |
    Horizontally scalable e-commerce API. All responses use an envelope:
    successful responses carry `data` (and `meta` for lists), errors carry
    `error: { code, message }`.

    Authentication is bearer JWT. Send `Idempotency-Key` on POST /orders and
    POST /payments to make retries safe.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                email: { type: string, format: email }
                phone: { type: string }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        "201": { $ref: "#/components/responses/AuthResult" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email or phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string }
                password: { type: string }
      responses:
        "200": { $ref: "#/components/responses/AuthResult" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token
      description: |
        Refresh tokens are single-use and belong to a family. Reusing a rotated
        token is treated as theft and revokes the whole family.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200": { $ref: "#/components/responses/AuthResult" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/password/forgot:
    post:
      tags: [Auth]
      summary: Request a password reset email
      description: Always returns success so accounts cannot be enumerated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: Accepted }
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset a password with a single-use token
      description: Revokes every existing session for the account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200": { description: Reset }
        "401": { $ref: "#/components/responses/Error" }
  /auth/password/change:
    post:
      tags: [Auth]
      summary: Change your password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200": { description: Changed }
        "401": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current tokens
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "204": { description: Revoked }
  /auth/me/export:
    get:
      tags: [Auth]
      summary: Export the caller's personal data (GDPR access)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: A JSON bundle of profile, addresses, orders, reviews and wishlist }
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Current profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/User" }
...
  /products:
    get:
      tags: [Catalog]
      summary: List or search products
      parameters:
        - { in: query, name: categoryId, schema: { type: string, format: uuid } }
        - { in: query, name: keyword, schema: { type: string } }
        - { in: query, name: sort, schema: { type: string, enum: [newest, price_asc, price_desc] } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: A page of products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Product" }
                      meta: { $ref: "#/components/schemas/Meta" }
  /products/{id}:
    get:
      tags: [Catalog]
      summary: Get a product
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The product
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/Error" }
  /cart:
    get:
      tags: [Cart]
      summary: Get the current cart
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Cart" }
    delete:
      tags: [Cart]
      summary: Clear the cart
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Cleared }
  /cart/items:
    post:
      tags: [Cart]
      summary: Add an item
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: string, format: uuid }
                quantity: { type: integer, minimum: 1, default: 1 }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "409": { $ref: "#/components/responses/Error" }
  /cart/items/{productId}:
    patch:
      tags: [Cart]
      summary: Set item quantity
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: productId, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 0 }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
    delete:
      tags: [Cart]
      summary: Remove an item
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: productId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
  /orders:
    post:
      tags: [Orders]
      summary: Checkout the cart
      description: Reserves stock atomically and creates a pending order.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: header, name: Idempotency-Key, schema: { type: string } }
      responses:
        "201":
          description: The order
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [Orders]
      summary: List orders (admins see all)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: status, schema: { $ref: "#/components/schemas/OrderStatus" } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: A page of orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Order" }
                      meta: { $ref: "#/components/schemas/Meta" }
  /orders/{id}/invoice:
    get:
      tags: [Orders]
      summary: Download the order invoice as a PDF
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: A PDF invoice
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "403": { description: Not the order owner }
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get an order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/Error" }
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order and release stock
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The cancelled order
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "409": { $ref: "#/components/responses/Error" }
  /payments:
    post:
      tags: [Payments]
      summary: Create a payment session
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: header, name: Idempotency-Key, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId]
              properties:
                orderId: { type: string, format: uuid }
                provider: { type: string, default: mock }
                returnUrl: { type: string, format: uri }
      responses:
        "201":
          description: The payment and redirect URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Payment" }
        "409": { $ref: "#/components/responses/Error" }
  /payments/simulate:
    post:
      tags: [Payments]
      summary: Sandbox — confirm a mock payment
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerRef]
              properties:
                providerRef: { type: string }
                provider: { type: string, default: mock }
      responses:
        "200": { description: Confirmed }
        "400": { $ref: "#/components/responses/Error" }
  /webhooks/payments/{provider}:
    post:
      tags: [Payments]
      summary: Provider callback (signature verified)
      parameters:
        - { in: path, name: provider, required: true, schema: { type: string } }
      responses:
        "200": { description: Accepted }
        "400": { $ref: "#/components/responses/Error" }
...
  /products/{id}/reviews:
    get:
      tags: [Catalog]
      summary: List reviews for a product
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: A page of reviews
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Review" }
                      meta: { $ref: "#/components/schemas/Meta" }
    post:
      tags: [Catalog]
      summary: Create a review (one per user per product)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                title: { type: string }
                body: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Review" }
        "409": { $ref: "#/components/responses/Error" }
  /reviews/{id}:
    patch:
      tags: [Catalog]
      summary: Update your review
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                title: { type: string }
                body: { type: string }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Catalog]
      summary: Delete a review (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
  /coupons/preview:
    post:
      tags: [Orders]
      summary: Validate a coupon against a subtotal
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                subtotalCents: { type: integer, format: int64 }
      responses:
        "200":
          description: The computed discount
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      code: { type: string }
                      discountCents: { type: integer, format: int64 }
                      totalCents: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /orders/{id}/returns:
    get:
      tags: [Orders]
      summary: List the caller's return requests for an order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Return requests }
    post:
      tags: [Orders]
      summary: Request a return for a delivered order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "201": { description: The return request }
        "409": { $ref: "#/components/responses/Error" }
...
  /addresses:
    get:
      tags: [Orders]
      summary: List the user's addresses
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Addresses }
    post:
      tags: [Orders]
      summary: Create an address
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddressInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Address" }
  /addresses/{id}:
    patch:
      tags: [Orders]
      summary: Update an address
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddressInput" }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Orders]
      summary: Delete an address
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
  /addresses/{id}/default:
    post:
      tags: [Orders]
      summary: Set the default address
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Updated }
  /orders/{id}/complete:
    post:
      tags: [Orders]
      summary: Confirm receipt of a shipped order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Completed }
        "409": { $ref: "#/components/responses/Error" }
...
  /auth/email/verify:
    post:
      tags: [Auth]
      summary: Verify an email with a single-use token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200": { description: Verified }
        "401": { $ref: "#/components/responses/Error" }
  /auth/email/resend:
    post:
      tags: [Auth]
      summary: Re-send the verification email
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Sent }
  /wishlist:
    get:
      tags: [Orders]
      summary: List saved products
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Products }
    post:
      tags: [Orders]
      summary: Save a product
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: string, format: uuid }
      responses:
        "201": { description: Saved }
  /wishlist/{productId}:
    delete:
      tags: [Orders]
      summary: Remove a saved product
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: productId, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Removed }
...
  /guest/orders/{token}:
    get:
      tags: [Orders]
      summary: View a guest order by access token
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200": { description: The order }
  /guest/orders/{token}/pay:
    post:
      tags: [Payments]
      summary: Create a payment for a guest order
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "201": { description: Payment session }
  /guest/orders/{token}/cancel:
    post:
      tags: [Orders]
      summary: Cancel a guest order
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200": { description: Cancelled }
  /guest/orders/{token}/complete:
    post:
      tags: [Orders]
      summary: Confirm receipt of a guest order
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200": { description: Completed }
  /guest/payments/simulate:
    post:
      tags: [Payments]
      summary: Sandbox confirm a guest payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerRef]
              properties:
                providerRef: { type: string }
                provider: { type: string }
      responses:
        "200": { description: Confirmed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    AuthResult:
      description: Authenticated session
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  expiresIn: { type: integer }
    Cart:
      description: The cart
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { $ref: "#/components/schemas/Cart" }
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
    Meta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        phone: { type: string }
        name: { type: string }
        role: { type: string, enum: [customer, admin] }
...
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        priceCents: { type: integer, format: int64 }
        currency: { type: string }
        coverImage: { type: string }
        images: { type: array, items: { type: string } }
        status: { type: string, enum: [draft, published, archived] }
        stock: { type: integer }
        weightGrams: { type: integer }
        variants:
          type: array
          items: { $ref: "#/components/schemas/Variant" }
        rating: { type: number }
        reviewCount: { type: integer }
...
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: string, format: uuid }
              variantId: { type: string, format: uuid }
              variantName: { type: string }
              title: { type: string }
              coverImage: { type: string }
              priceCents: { type: integer, format: int64 }
              currency: { type: string }
              quantity: { type: integer }
        totalCents: { type: integer, format: int64 }
        totalCount: { type: integer }
    OrderStatus:
      type: string
      enum: [pending_payment, paid, cancelled, shipped, completed, refunded]
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderNo: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        currency: { type: string }
        subtotalCents: { type: integer, format: int64 }
        discountCents: { type: integer, format: int64 }
        couponCode: { type: string }
        shippingCents: { type: integer, format: int64 }
        taxCents: { type: integer, format: int64 }
        shippingMethod: { type: string }
        totalCents: { type: integer, format: int64 }
        paymentId: { type: string }
        shippingAddress: { $ref: "#/components/schemas/Address" }
        trackingNo: { type: string }
        shippedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        paidAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              productId: { type: string, format: uuid }
              variantId: { type: string, format: uuid }
              variantName: { type: string }
              sku: { type: string }
              title: { type: string }
              priceCents: { type: integer, format: int64 }
              quantity: { type: integer }
              subtotal: { type: integer, format: int64 }
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        provider: { type: string }
        status: { type: string, enum: [pending, succeeded, failed, refunded] }
        amountCents: { type: integer, format: int64 }
        currency: { type: string }
        redirectUrl: { type: string, format: uri }
...
    AddressInput:
      type: object
      required: [recipient, line1]
      properties:
        recipient: { type: string }
        phone: { type: string }
        province: { type: string }
        city: { type: string }
        district: { type: string }
        line1: { type: string }
        postalCode: { type: string }
        default: { type: boolean }
    Address:
      type: object
      properties:
        id: { type: string, format: uuid }
        recipient: { type: string }
        phone: { type: string }
        province: { type: string }
        city: { type: string }
        district: { type: string }
        line1: { type: string }
        postalCode: { type: string }
        default: { type: boolean }
...
    Review:
      type: object
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        title: { type: string }
        body: { type: string }
        verifiedPurchase: { type: boolean, description: "True when the reviewer has a paid order for the product" }
        createdAt: { type: string, format: date-time }
...